+++ src/remote_generator/services/atom_entity_generator.rb
@@ -15,7 +15,7 @@ module Foobara
         end
 
         def target_path
-          [*domain.scoped_full_path, "types", entity_name, "Atom.ts"]
+          [*super[..-2], "Atom.ts"]
         end
 
         def template_path