+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-## [unreleased] - 2024-08-2x
+## [0.0.4] - 2024-08-22
 
 - Add type generators to command result generator
 - Add type generators to domain generator
+- Attempt to fix botched nested type paths
 
 ## [0.0.2] - 2024-08-21